requirements.txt
mkdocs
pip
pytest
python-dotenv
ruff
torch>=2.0.0
torchvision>=0.15.0
transformers[torch]>=4.30.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.5.0
pandas>=1.5.0
opencv-python>=4.7.0
evaluate>=0.4.0
jiwer>=3.0.0
tqdm>=4.64.0
